refactorings to formatting
