+++ server/src/main/kotlin/org/javacs/kt/formatting/NopFormatter.kt
@@ -1,0 +1,8 @@
+package org.javacs.kt.formatting
+
+import org.eclipse.lsp4j.FormattingOptions as LspFormattingOptions
+
+object NopFormatter : Formatter {
+    override fun format(code: String, options: LspFormattingOptions): String = code
+}
+
